Adding convert methods for client_max_body_size

+++ srcs/config/ConfigHelper.hpp
@@ -7,5 +7,5 @@ class ConfigHelper {
 	public:
 		static bool convertStringToBool(std::string const &str);
 		static int convertStringToPort(std::string const &str);
-		static int convertStringToClientMaxBodySize(std::string const &str);
+		static double convertStringToClientMaxBodySize(std::string const &str);
 };
